bugfix: make layout static instead of dynamic
