+++ src/lib/auth.ts
@@ -1,6 +1,10 @@
+import 'server-only';
+
 import { betterAuth } from 'better-auth';
 import { drizzleAdapter } from 'better-auth/adapters/drizzle';
+import { headers } from 'next/headers';
 import { db } from '~/server/db';
+import { nextCookies } from 'better-auth/next-js';
 
 export const auth = betterAuth({
   database: drizzleAdapter(db, {
@@ -9,6 +13,13 @@ export const auth = betterAuth({
   emailAndPassword: {
     enabled: true,
   },
+  plugins: [nextCookies()],
 });
+
+export async function getSession() {
+  return await auth.api.getSession({
+    headers: await headers(),
+  });
+}
 
 export type Session = Awaited<ReturnType<typeof auth.api.getSession>>;
