+++ src/stores/chat.store.ts
@@ -1,3 +1,5 @@
+import 'client-only';
+
 import { create } from 'zustand';
 import { useShallow } from 'zustand/react/shallow';
 import type { Document } from '~/server/db/schema';
